Split renderer base classes into modules
