+++ src/heart/renderers/__init__.py
@@ -1,3 +1,3 @@
-from .base import AtomicBaseRenderer as AtomicBaseRenderer
+from .atomic import AtomicBaseRenderer as AtomicBaseRenderer
 from .base import BaseRenderer as BaseRenderer
-from .base import StatefulBaseRenderer as StatefulBaseRenderer
+from .stateful import StatefulBaseRenderer as StatefulBaseRenderer
